requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.19.0
python-dotenv==1.0.1
email-validator==2.2.0
openai==1.91.0
fastapi==0.116.1
uvicorn==0.35.0